fix(rsc): isolate views counter server fn from client bundle
